requirements.txt
alpaca-trade-api
scikit-learn
pandas
numpy
matplotlib
streamlit
python-dotenv
